+++ integration-tests/require-after-teardown/__tests__/late-require.test.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+'use strict';
+
+test('require after done', () => {
+  setTimeout(() => {
+    const double = require('../');
+
+    expect(double(5)).toBe(10);
+  }, 0);
+});