+++ packages/expect/src/__tests__/matchers.test.js
@@ -204,6 +204,34 @@ describe('.toBe()', () => {
   });
 });
 
+describe('.toStrictEqual()', () => {
+  class TestClass {
+    constructor(a, b) {
+      this.a = a;
+      this.b = b;
+    }
+  }
+
+  it('does not ignore keys with undefined values', () => {
+    expect({
+      a: undefined,
+      b: 2,
+    }).not.toStrictEqual({b: 2});
+  });
+
+  it('passes when comparing same type', () => {
+    expect({
+      test: new TestClass(1, 2),
+    }).toStrictEqual({test: new TestClass(1, 2)});
+  });
+
+  it('does not pass for different types', () => {
+    expect({
+      test: new TestClass(1, 2),
+    }).not.toStrictEqual({test: {a: 1, b: 2}});
+  });
+});
+
 describe('.toEqual()', () => {
   [
     [true, false],