+++ packages/jest-config/src/__tests__/defaults.test.js
@@ -1,0 +1,5 @@
+import {defaults} from '../index';
+
+test('get configuration defaults', () => {
+  expect(defaults).toBeDefined();
+});