+++ packages/jest-editor-support/index.d.ts
@@ -13,7 +13,6 @@ export interface SpawnOptions {
 }
 
 export interface Options {
-  coverage?: boolean;
   createProcess?(
     workspace: ProjectWorkspace,
     args: string[],
@@ -49,11 +48,13 @@ export class ProjectWorkspace {
     pathToJest: string,
     pathToConfig: string,
     localJestMajorVersin: number,
+    collectCoverage?: boolean,
   );
   pathToJest: string;
   pathToConfig: string;
   rootPath: string;
   localJestMajorVersion: number;
+  collectCoverage?: boolean;
 }
 
 export interface IParseResults {