+++ packages/jest-haste-map/src/index.js
@@ -542,6 +542,7 @@ class HasteMap extends EventEmitter {
   _cleanup() {
     const worker = this._worker;
 
+    // $FlowFixMe
     if (worker && typeof worker.end === 'function') {
       worker.end();
     }
@@ -646,7 +647,9 @@ class HasteMap extends EventEmitter {
     const Watcher =
       canUseWatchman && this._options.useWatchman
         ? WatchmanWatcher
-        : os.platform() === 'darwin' ? sane.FSEventsWatcher : sane.NodeWatcher;
+        : os.platform() === 'darwin'
+          ? sane.FSEventsWatcher
+          : sane.NodeWatcher;
     const extensions = this._options.extensions;
     const ignorePattern = this._options.ignorePattern;
     let changeQueue = Promise.resolve();
@@ -724,7 +727,6 @@ class HasteMap extends EventEmitter {
 
           if (mustCopy) {
             mustCopy = false;
-            // $FlowFixMe
             hasteMap = {
               clocks: copy(hasteMap.clocks),
               duplicates: copy(hasteMap.duplicates),
@@ -753,7 +755,6 @@ class HasteMap extends EventEmitter {
             if (Object.keys(moduleMap).length === 0) {
               delete hasteMap.map[moduleName];
             } else {
-              // $FlowFixMe
               hasteMap.map[moduleName] = moduleMap;
             }
           }