+++ packages/jest-jasmine2/src/jasmine/Env.js
@@ -32,6 +32,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import queueRunner from '../queue_runner';
 import treeProcessor from '../tree_processor';
+import prettyFormat from 'pretty-format';
 
 // Try getting the real promise object from the context, if available. Someone
 // could have overridden it in a test. Async functions return it implicitly.
@@ -547,11 +548,12 @@ export default function(j$) {
     };
 
     this.fail = function(error) {
-      let message = 'Failed';
-      if (error) {
-        message += ': ';
-        message += error.message || error;
-      }
+      // duck-type Error, see #2549
+      const isError =
+        typeof error === 'object' &&
+        typeof error.message === 'string' &&
+        typeof error.name === 'string';
+      const message = `Failed: ${prettyFormat(error, {maxDepth: 3})}`;
 
       currentRunnable().addExpectationResult(false, {
         matcherName: '',
@@ -559,7 +561,7 @@ export default function(j$) {
         expected: '',
         actual: '',
         message,
-        error: error && error.message ? error : null,
+        error: isError ? error : new Error(message),
       });
     };
   }