+++ packages/jest-mock/src/__tests__/jest_mock.test.js
@@ -508,67 +508,64 @@ describe('moduleMocker', () => {
       expect(fn.mock.thrownErrors).toEqual([undefined, error, undefined]);
     });
 
-    describe('timestamps', () => {
-      const RealDate = Date;
+    describe('invocationCallOrder', () => {
+      it('tracks invocationCallOrder made by mocks', () => {
+        const fn1 = moduleMocker.fn();
+        expect(fn1.mock.invocationCallOrder).toEqual([]);
 
-      beforeEach(() => {
-        global.Date = {
-          now: jest
-            .fn()
-            .mockImplementationOnce(() => 978391040765)
-            .mockImplementationOnce(() => 1262388620765),
-        };
+        fn1(1, 2, 3);
+        expect(fn1.mock.invocationCallOrder[0]).toBe(1);
+
+        fn1('a', 'b', 'c');
+        expect(fn1.mock.invocationCallOrder[1]).toBe(2);
+
+        fn1(1, 2, 3);
+        expect(fn1.mock.invocationCallOrder[2]).toBe(3);
+
+        const fn2 = moduleMocker.fn();
+        expect(fn2.mock.invocationCallOrder).toEqual([]);
+
+        fn2('d', 'e', 'f');
+        expect(fn2.mock.invocationCallOrder[0]).toBe(4);
+
+        fn2(4, 5, 6);
+        expect(fn2.mock.invocationCallOrder[1]).toBe(5);
       });
 
-      afterEach(() => {
-        global.Date = RealDate;
-      });
-
-      it('tracks timestamps made by mocks', () => {
+      it('supports clearing mock invocationCallOrder', () => {
         const fn = moduleMocker.fn();
-        expect(fn.mock.timestamps).toEqual([]);
+        expect(fn.mock.invocationCallOrder).toEqual([]);
 
         fn(1, 2, 3);
-        expect(fn.mock.timestamps[0]).toBe(978391040765);
-
-        fn('a', 'b', 'c');
-        expect(fn.mock.timestamps[1]).toBe(1262388620765);
-      });
-
-      it('supports clearing mock timestamps', () => {
-        const fn = moduleMocker.fn();
-        expect(fn.mock.timestamps).toEqual([]);
-
-        fn(1, 2, 3);
-        expect(fn.mock.timestamps).toEqual([978391040765]);
+        expect(fn.mock.invocationCallOrder).toEqual([1]);
 
         fn.mockReturnValue('abcd');
 
         fn.mockClear();
-        expect(fn.mock.timestamps).toEqual([]);
+        expect(fn.mock.invocationCallOrder).toEqual([]);
 
         fn('a', 'b', 'c');
-        expect(fn.mock.timestamps).toEqual([1262388620765]);
+        expect(fn.mock.invocationCallOrder).toEqual([2]);
 
         expect(fn()).toEqual('abcd');
       });
 
-      it('supports clearing all mocks timestamps', () => {
+      it('supports clearing all mocks invocationCallOrder', () => {
         const fn1 = moduleMocker.fn();
         fn1.mockImplementation(() => 'abcd');
 
         fn1(1, 2, 3);
-        expect(fn1.mock.timestamps).toEqual([978391040765]);
+        expect(fn1.mock.invocationCallOrder).toEqual([1]);
 
         const fn2 = moduleMocker.fn();
 
         fn2.mockReturnValue('abcde');
         fn2('a', 'b', 'c', 'd');
-        expect(fn2.mock.timestamps).toEqual([1262388620765]);
+        expect(fn2.mock.invocationCallOrder).toEqual([2]);
 
         moduleMocker.clearAllMocks();
-        expect(fn1.mock.timestamps).toEqual([]);
-        expect(fn2.mock.timestamps).toEqual([]);
+        expect(fn1.mock.invocationCallOrder).toEqual([]);
+        expect(fn2.mock.invocationCallOrder).toEqual([]);
         expect(fn1()).toEqual('abcd');
         expect(fn2()).toEqual('abcde');
       });