+++ packages/jest-mock/src/index.js
@@ -26,7 +26,7 @@ type MockFunctionState = {
   calls: Array<Array<any>>,
   returnValues: Array<any>,
   thrownErrors: Array<any>,
-  timestamps: Array<number>,
+  invocationCallOrder: Array<number>,
 };
 
 type MockFunctionConfig = {
@@ -161,7 +161,11 @@ function isA(typeName: string, value: any): boolean {
 }
 
 function getType(ref?: any): string | null {
-  if (isA('Function', ref) || isA('AsyncFunction', ref)) {
+  if (
+    isA('Function', ref) ||
+    isA('AsyncFunction', ref) ||
+    isA('GeneratorFunction', ref)
+  ) {
     return 'function';
   } else if (Array.isArray(ref)) {
     return 'array';
@@ -194,7 +198,9 @@ function isReadonlyProp(object: any, prop: string): boolean {
       prop === 'callee' ||
       prop === 'name' ||
       prop === 'length') &&
-      (isA('Function', object) || isA('AsyncFunction', object))) ||
+      (isA('Function', object) ||
+        isA('AsyncFunction', object) ||
+        isA('GeneratorFunction', object))) ||
     ((prop === 'source' ||
       prop === 'global' ||
       prop === 'ignoreCase' ||
@@ -219,7 +225,7 @@ function getSlots(object?: Object): Array<string> {
       const prop = ownNames[i];
       if (!isReadonlyProp(object, prop)) {
         const propDesc = Object.getOwnPropertyDescriptor(object, prop);
-        if (!propDesc.get || object.__esModule) {
+        if ((propDesc !== undefined && !propDesc.get) || object.__esModule) {
           slots[prop] = true;
         }
       }
@@ -235,6 +241,7 @@ class ModuleMockerClass {
   _mockConfigRegistry: WeakMap<Function, MockFunctionConfig>;
   _spyState: Set<() => void>;
   ModuleMocker: Class<ModuleMockerClass>;
+  _invocationCallCounter: number;
 
   /**
    * @see README.md
@@ -247,6 +254,7 @@ class ModuleMockerClass {
     this._mockConfigRegistry = new WeakMap();
     this._spyState = new Set();
     this.ModuleMocker = ModuleMockerClass;
+    this._invocationCallCounter = 1;
   }
 
   _ensureMockConfig(f: Mock): MockFunctionConfig {
@@ -282,9 +290,9 @@ class ModuleMockerClass {
     return {
       calls: [],
       instances: [],
+      invocationCallOrder: [],
       returnValues: [],
       thrownErrors: [],
-      timestamps: [],
     };
   }
 
@@ -319,7 +327,7 @@ class ModuleMockerClass {
         const mockConfig = mocker._ensureMockConfig(f);
         mockState.instances.push(this);
         mockState.calls.push(Array.prototype.slice.call(arguments));
-        mockState.timestamps.push(Date.now());
+        mockState.invocationCallOrder.push(mocker._invocationCallCounter++);
 
         // Will be set to the return value of the mock if an error is not thrown
         let finalReturnValue;
@@ -793,7 +801,9 @@ class ModuleMockerClass {
       }
 
       descriptor[accessType] = this._makeComponent({type: 'function'}, () => {
+        // $FlowFixMe
         descriptor[accessType] = original;
+        // $FlowFixMe
         Object.defineProperty(obj, propertyName, descriptor);
       });