+++ packages/jest-runtime/src/__tests__/script_transformer.test.js
@@ -85,6 +85,44 @@ jest.mock(
   {virtual: true},
 );
 
+// Bad preprocessor
+jest.mock(
+  'skipped-required-props-preprocessor',
+  () => {
+    return {};
+  },
+  {virtual: true},
+);
+
+// Bad preprocessor
+jest.mock(
+  'skipped-required-create-transformer-props-preprocessor',
+  () => {
+    return {
+      createTransformer() {
+        return {};
+      },
+    };
+  },
+  {virtual: true},
+);
+
+jest.mock(
+  'skipped-process-method-preprocessor',
+  () => {
+    return {
+      createTransformer() {
+        const mockProcess = jest.fn();
+        mockProcess.mockReturnValue('code');
+        return {
+          process: mockProcess,
+        };
+      },
+    };
+  },
+  {virtual: true},
+);
+
 const getCachePath = (fs, config) => {
   for (const path in mockFs) {
     if (path.startsWith(config.cacheDirectory)) {
@@ -261,6 +299,38 @@ describe('ScriptTransformer', () => {
       });
     },
   );
+
+  it("throws an error if `process` doesn't defined", () => {
+    Object.assign(config, {
+      transform: [['^.+\\.js$', 'skipped-required-props-preprocessor']],
+    });
+    const scriptTransformer = new ScriptTransformer(config);
+    expect(() =>
+      scriptTransformer.transformSource('sample.js', '', false),
+    ).toThrow('Jest: a transform must export a `process` function.');
+  });
+
+  it('throws an error if createTransformer returns object without `process` method', () => {
+    Object.assign(config, {
+      transform: [
+        ['^.+\\.js$', 'skipped-required-create-transformer-props-preprocessor'],
+      ],
+    });
+    const scriptTransformer = new ScriptTransformer(config);
+    expect(() =>
+      scriptTransformer.transformSource('sample.js', '', false),
+    ).toThrow('Jest: a transform must export a `process` function.');
+  });
+
+  it("shouldn't throw error without process method. But with corrent createTransformer method", () => {
+    Object.assign(config, {
+      transform: [['^.+\\.js$', 'skipped-process-method-preprocessor']],
+    });
+    const scriptTransformer = new ScriptTransformer(config);
+    expect(() =>
+      scriptTransformer.transformSource('sample.js', '', false),
+    ).not.toThrow();
+  });
 
   it('uses the supplied preprocessor', () => {
     config = Object.assign(config, {